requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==6.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.11
marshmallow==3.26.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
transbank-sdk==6.1.0
urllib3==2.6.2
whitenoise==6.11.0
